Add i18n support
Additional locales can be contributed to the `/resources/i18n` folder.
The pattern is `coda_{LANG}(_{LOCALE}).json` with `{LOCALE}` being optional.
cbi18n will use the most specific file first and fall back if not.
English is used as a global fallback if a key is not found.

Tests can be added by duplicating the `FormatDistance_EN_US_Spec.cfc`,
renaming it to match your new locale, and changing the functions to
return the correct translation.  The test will be ran automatically for you.